moved over gap certification functions to a new file
